Create properties app with BoardingHouse model
The commit moves the BoardingHouse model from utils to a new dedicated
properties app, creating a more organized structure for property-related
features.
