+++ utils/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from cloudinary.models import CloudinaryField
 
 from django.contrib.contenttypes.models import ContentType
-from django.contrib.contenttypes.fields import GenericForeignKey, GenericRelation
+from django.contrib.contenttypes.fields import GenericForeignKey
 
 from accounts.utils import UserType
 
@@ -40,23 +40,6 @@ class ImageRelation(models.Model):
 
     class Meta:
         db_table = "image_relation"
-
-
-class BoardingHouse(BaseModel):
-    """
-    Model này định nghĩa một dãy trọ:
-    Khi chủ trọ mới tạo một tài khoản, bắt buộc chủ trọ phải tạo một dãy trọ,
-    và dãy trọ này sẽ được xét duyệt bởi quản trị viên. Nếu dãy trọ đầu tiên của
-    chủ trọ đã không được xét duyệt, thì tài khoản của chủ trọ sẽ không được phép
-    tạo mới.
-
-    Ngược lại, nếu dãy trọ đã được xét duyệt, thì chủ trọ có thể tạo các bài đăng
-    """
-
-    boarding_house_name = models.CharField(max_length=256)
-    address = models.CharField(max_length=256)
-    images = GenericRelation(Image, null=True)
-
 
 
 class CommentPost(BaseModel):
